Moved global.setup.ts and global.teardown.ts into tests/ dir
