+++ e2e/tests/global.setup.ts
@@ -1,5 +1,5 @@
 import {test as setup} from '@playwright/test';
-import {EnvironmentManager} from './helpers/environment/EnvironmentManager';
+import {EnvironmentManager} from '../helpers/environment';
 
 setup('global environment setup', async () => {
     const environmentManager = new EnvironmentManager();
