+++ e2e/tests/global.teardown.ts
@@ -1,7 +1,7 @@
 import {test as teardown} from '@playwright/test';
-import {EnvironmentManager} from './helpers/environment/EnvironmentManager';
+import {EnvironmentManager} from '../helpers/environment';
 
-teardown('global environment cleanup', async ({}) => {
+teardown('global environment cleanup', async () => {
     const environmentManager = new EnvironmentManager();
     await environmentManager.globalTeardown();
 });
